// client/src/constants/dataDefault.ts
export interface Info {
    StoreI: string;
    StoreII: string;
    City: string;
    Email: string;
    Phone: string;
    LinkFacebook: string;
    LinkInstagram: string;
    LinkTwitter: string;
    LinkYoutobe: string;
    NameLogo: string;
}
export const defaultInfo: Info = {
    StoreI: "50 Phạm Văn Đồng, Gò Vấp, TP. HCM",
    StoreII: "81 Phạm Ngủ Lão, Gò Vấp, TP. HCM",
    City: "TP. HCM, Việt Nam",
    Email: "tuannguyenipv6@gmail.com",
    Phone: "0987654321",
    LinkFacebook: "https://www.facebook.com/NQT.TuanNguyen.ipv6",
    LinkInstagram: "https://www.instagram.com/NQT.TuanNguyen.ipv6/",
    LinkTwitter: "https://twitter.com/",
    LinkYoutobe: "https://www.youtube.com/channel/UCv32aGaK67dxxGtieStdWeQ",
    NameLogo: "ShopNQT.png"
}